requirements.txt
aiohttp==3.9.5  # Replace requests with aiohttp for async HTTP requests
alembic==1.14.0
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
nest_asyncio==1.6.0  # Added for async support in Flask
numpy==2.2.1
pillow==11.1.0  # For QR code scanning with PIL
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzbar==0.1.9  # For QR code decoding
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0  # Still needed as a dependency for aiohttp
Werkzeug==3.1.3